move object data definition types to separate modules
- update e_object_team names
Auto stash before rebase of "refs/heads/objects"
