+++ xlive/Blam/Engine/ai/actors.h
@@ -5,7 +5,8 @@
 
 #include "Blam/Cache/DataTypes/BlamDataTypes.h"
 #include "Blam/Engine/memory/data.h"
-#include "Blam/Engine/scenario/scenario.h"
+#include "Blam/Engine/objects/objects.h"
+#include "Blam/Engine/objects/object_placement.h"`
 #include "Blam/Math/real_math.h"
 
 /*
