+++ xlive/Blam/Engine/items/items.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "Blam/Engine/objects/objects.h"
+
+struct s_item_data_definition
+{
+	s_object_data_definition object;
+	char field_0[64];
+};
+CHECK_STRUCT_SIZE(s_item_data_definition, 364);
