+++ xlive/Blam/Engine/units/bipeds.h
@@ -1,0 +1,23 @@
+#pragma once
+#include "units.h"
+#include "Blam/Engine/tag_files/string_id.h"
+
+enum e_biped_physics_mode : byte
+{
+	mode_ground = 1,
+	mode_flying,
+	mode_dead,
+	mode_posture,
+	mode_climbing,
+	mode_melee
+};
+struct s_biped_data_definition : s_unit_data_definition
+{
+	PAD(0x3F4 - sizeof(s_unit_data_definition));
+	e_biped_physics_mode biped_mode;//0x3F4
+	PAD(0x480 - 0x3F5);
+
+	// NEW DATA
+	string_id variant_name;
+};
+CHECK_STRUCT_SIZE(s_biped_data_definition, 0x480 + 4);
