+++ xlive/H2MOD/GUI/XLiveRendering.cpp
@@ -5,6 +5,7 @@
 #include "backends/imgui_impl_dx9.h"
 #include "backends/imgui_impl_win32.h"
 
+#include "Blam/Engine/units/bipeds.h"
 #include "Blam/Engine/game/game.h"
 
 #include "H2MOD/Modules/Accounts/AccountLogin.h"
@@ -12,7 +13,6 @@
 #include "H2MOD/Modules/Achievements/Achievements.h"
 #include "H2MOD/Modules/Shell/Config.h"
 #include "H2MOD/Modules/Input/KeyboardInput.h"
-#include "H2MOD/Modules/Input/PlayerControl.h"
 #include "H2MOD/Modules/Networking/Networking.h"
 #include "H2MOD/Modules/OnScreenDebug/OnscreenDebug.h"
 #include "H2MOD/Modules/Shell/Shell.h"
